Use context to send additional metadata to Sentry
Previously we were using tags, which isn't quite right. They are
searchable but only support primitives as a result, which didn't make
sense for the data we wanted to send.

This instead moves toward using context, which is more readable in the
Sentry interface and allows us to send complex data like we were before.
More of a drop-in replacement for how we were using Errbit.

For this approach, we just need to watch the size of the payload. We
could set additional contexts with unique names in the future. More
on that here: https://develop.sentry.dev/sdk/event-payloads/contexts/

CV2-2897