Add ApiKey as user information in Sentry
This will add an API Key ID, similar to what we use in Honeycomb,
to error information sent to Sentry.

This uses the same interface as Check API's Sentry class, but sets
the user_id as the API key since that is our closest proxy for a
user in this application. Check API uses the actual User ID for this.

CV2-2897