Implement a FileItem parser
Currently we're causing an error when a user submits a file URL
to Pender, either via the Check web interface or API. This means that
partners can't submit links pointing to files.

While we want to eventually detect filetype using a HEAD request to the
URL before fetching the file itself, this is a quick and easy way
for us to do a best guess at most file types we would expect using the
current flow - making a get request, then passing it off to a parser
that matches the URL structure.

In order to prevent erroring on the Check API side, I also had to
prevent the oembed from returning the file data. This invovled what
I hope is a temporary carveout for file parsers, and telling them
to use the default Oembed.

The result of all of this is that users can submit a link to a file
and just see a URL with no additional metadata or oembed in the
Check view, kind of like our normal fallback behavior.

CV2-2724