change get_jsonld_data to receive and return an array
Note:
Commented media.add_schema_to_data as we are very likely to delete it.
Seems like it was a way to answer annotation fields automatically based
on JSON+LD data, but we haven't been using it. And right now it wouldn’t
function well anyway, since we only support one schema and a document
will likely have multiple