fix failing archiver test
Our test was failing because the https://edition.cnn.com url now redirects
to cnn.com, which meant we were getting two different ids.

We changed the test to stub a request instead of relying on a live url,
which could change again at any time and cause this test to fail.