remove reddit specific test
This is the file for the general page parser, we don't think an
page specific test should be here.

We are assuming there was undesirable behavior for reddit at some point
so we added an integration test to try to prevent against regressions.
but without any more context or a unit test, it’s hard to know much more.
We don’t think it’s a useful test- hard to be 100% sure, but we think we
can and should delete