Attept to address test flakiness
Sometimes our tests fail when run in a certain order. For an example
try bin/rails test test/models/parser/facebook_item_test.rb --seed 8782
before this commit.

I believe this is due to us using the same URL across integration tests
and unit tests -- if the integration tests are run first then the URL
is saved in the redis cache, and our integration tests look to that
rather than attempting to make a request against our fake crowdtangle.
As a result, they never hit the test data.

This commit changes the URLs that the unit tests use to be something
unique and fake, and also moves the fixture files directly into the
tests because we don't use them anywhere else. I'm honestly kind of
ambivalent about that but I had already done the work while trying
to troubleshoot the tests and don't want to undo it. Real talk.