CV2-3119: protection weakened or disabled warning (#355)
* remove protect_from_forgery and keep skip_before_action

When we add protect_from_forgery and then skip_before_action, we make the first one pointless, and it can cause confusion since it doesn't have a real reason to be there. Since this is a stateless API we don't need to use protect_from_forgery, so we can only keep skip_before_action.