update guard clause to avoid undefined method error (#363)
The problem here was we were looking for a key that didn't exist in the
params hash. If we get an url with a query for photos, like
"https://www.facebook.com/search/photos?q=%E0%...". It would look for
the values from the key "album_id" in the params hash, but that key
does not exist, the only key there is "q", and that returns a query,
not the id.

So I updated the guard clause to check if the key exists in params.