3019 - add safe operator to avoid NoMethodError (kwai) (#364)
* add safe operator

* removed test that checks for NoMethodError

This test only worked when the two selectors we look for don't exist,
but sometimes a link might not have those specific selectors and still
have relevant information to be returned.

By adding a safe operator we are not getting the NoMethodError anymore,
and are returning relevant data. But this test stops making sense.
We thought about adding tests to specific links or errors, but in the end
we think those are not necessary.