small fixes to get method
It wasn't working before because of an extra '/' in the path, we removed
that, and updated to use URI.join, which should help avoid that kind of issue.

Updated to use URI insetad of our parse_url helper. We know the uri that
is coming, so we don't need all that, URI is sufficient. Also use_ssl can
just be true here.