refactor twitter_author_url method
I had just substituted for the new class, but the class is now
just a concern like the other providers.

Besides that the url the new api returns is the url linked
in the profile, not the twitter url. So for now I'm only combining
the twitter url with the username to get the author url.

Also removed the error handling since it was the error handling from
the twitter gem, and we are not making an api call to get this info.