Update base.rb and it's test usage of the twitter api/gem
Yesterday these tests broke because I updated the method we were
using from the Twitter gem in the base parser, but I had forgotten to
update it once I changed the new class to a module.

Since I fixed that, tests seem to be working. (They were successful locally)

I removed functionallity that was related to the Twitter gem from the
tests, but left comments in two tests that I need to take a better look.
One of those I'm skipping, even though it works, because I don't think
it makes sense anymore, but probably should be adapted.