upgrade postgres image to 13
I had upgraded to postgres12-bullseye because of a issue we had when
building on Travis:
We had an issue with building on Travis that seems related to a change
the maintainers of the postgres docker images have made to the underlying
OS image layer: Previous: Debian 11 (bullseye) New: Debian 12 (bookworm).
The workaround seems to be using postgres-bullseye.More on this here
https://stackoverflow.com/questions/76555305/postgres-container-failed-to-start-with-initdb-error-popen-failure-cannot-allo/76591040#76591040

Now updating to 13 I checked which image Devin used in Alegre and am
using the same one here.