update according to Christa's review
main notes:

- instead of re-raising the error inside the provider, we are notifying
sentry and returning an errors hash.
https://github.com/meedan/pender/pull/371#discussion_r1290459001

- we re-wrote the parsers according to what we feel is 'safer' moving
forward: using merge! to set the defaults.

- I had to update some of page_item_tests, so I used this as an opportuniy
to move the integration tests to their own file. (we are working on moving
all the integration tests to their own files, separated from the unit ones)

- Updated the error tests, now we test 3 scenarios:
200 response with error in json, non-200 response, exception in Net::HTTP

- Updated to use dig, ie. parsed_data.dig('raw','api','data',0). this will
return nil if data is missing, before it would raise an error for invalid access