update unavailable pages tests (#374)
Sometimes facebook items are unavailable. Before we didn’t get any
response in those cases, now we get a page letting us know the
page is not available. With this we are now able to create an embed.

We want to keep this new behaviour, so we only need to update
the tests.