keep (function(){...})()
The form (function(){...})() is a function literal that is executed
immediately. What this means is that you have defined a function and
you are calling it immediately.

This form is useful for information hiding and encapsulation since
anything you define inside that function remains local to that function
and inaccessible from the outside world.