3227 - Remove vendorized jquery (#375)
* remove jquery

We were only using jquery in the css block that was deleted.
relevant PR: https://github.com/meedan/pender/pull/368

We kept (function(){...})(). This form is useful for information hiding and encapsulation since anything you define inside that function remains local to that function and inaccessible from the outside world.