make sure oembed_json is a Hash (#376)
We sometimes get an error when parsing the oembed: TypeError: no implicit conversion of String into Integer

This happens with similar specific urls, all weirdly named urls with .top domain. Those sites return '404' on their response.body, when we JSON.parse it, it returns an integer, which breaks things.

This was fixed by returning early if the value is not a Hash.