update story regex and update facebook_item parser
We don't want the title and description to be set to 'log in to facebook'.
The title should be the url and description should be an empty string.

We were having two issues:
1. the story link was being parsed by the page parser, so we updated the
regex
2. the facebook parser was being redirected to a login page, which meant
that the metadata we used for the title and description was 'log in...'.
now we check one final time and update the title if needed.