3653 - extract username correctly when there is a query (#385)
* add regex to the twitter profile parser

sometimes a profile url might come with a query, we need to make sure
we are getting only the username to make the request to the twitter api
endpoint

At first I tried to fit everything into two patterns, which ended up
matching the twitter item as well. So I broke it into two more patterns,
and ordered them from the two that are more specific to the less.