move integration tests to unit
- test "should get canonical URL from facebook object 3"
- test "should create Facebook post from mobile URL"
- test "should get canonical URL from facebook object 3"
refactored to make it clearer

- test "should not use Facebook embed if is a link to redirect"
moved it, but it has issues:
1. it's hiting profile not item
2. this link doesn't redirect from FB to a different site, it hits a
'post doesn't exist' sort of message
3. I tried looking for a post that did redirect but I think that
behavior might no longer exist

So I'm not quite sure what we are testing, if this test is needed, and if
we should update the regex so this stuff will hit item. I'll leave this
here for now and come back to it to decide what to do with this

- test "should return canonical url when redirected to login page"
this used to be "should not change url when redirected to login page" but
it had been failing because it now returns the canonical_url. I'm not
sure why this changed, but I'm not sure it's an issue

- test "should add login required error, return html and empty description"
I think we have some repetitive testing going on, after I finish moving all
of them I'm going to lokk through what we have, refactor and clean up

- test "should get canonical URL parsed from facebook html when it is relative"
- test "should get canonical URL parsed from facebook html when it is a page"
- test "should get the group name when parsing group post"
- test "should return html even when FB url is private"
- test "should store oembed data of a facebook post"