add test and try to fix failing one
The url fallback is handled by Media. I wanted to test that, and the Kwai
parser directly, so I made two tests:
test "assigns description and username to hash from the json+ld"
"assigns values to hash from the json+ld and falls back to url as title"

The second test had worked locally, but failed on Travis, then failed in
weird ways locally, but now seems to be successful. I'll run it on Travis
again and see how this behaves.