3723 – Fix Kwai parser (#387)
* update kwai parser to use ldjson

The parser stopped working: we search the doc for '.info .title' and for '.name', and both of those are returning nil.

Kwai has changed their HTML and changed it back again before. So we think it's worth keeping the old behavior, at least for now. We are getting data from the jsonld script as fallback. 

We get the name from creator name, use description or  transcription for data['description']. For title we don't have a good substitute, so we are just using the default fallback from Media (url).

* add and update tests

The url fallback is handled by Media. I wanted to test that, and the Kwai
parser directly, so I made two tests:
- test "assigns description and username to hash from the json+ld"
- test "assigns values to hash from the json+ld and falls back to url as title"