clean up FacebookItemUnitTest
- we didn't need to store this parser.parse_data(empty_doc, throwaway_url)
inside a var, since it's not being called (we do need this to run for
the test to work)
- marked a test "sends error to sentry when we receive unexpected response
from crowdtangle API" as flaky
     - changed the flaky test .to_return(status: 200, body: '') to
     .to_return(status: 200, body: 'something unexpected') just to make it
     clearer. this doesn't address the flakyness
- updated test "should create Facebook item from mobile URL"
I thought it could be clearer if really the only thing we are making sure
is that this parses mobile links. We are not able to check for the provider
though, as that is set in Media. However, if we are testing for the patterns,
ans testing for the parsing behavior already, is this test necessary?
- decided to delete test "should not use Facebook embed if is a link to redirect"
     It had a few issues:
     - 1. it's hiting profile not item
     - 2. this link doesn't redirect from FB to a different site, it hits a 'post doesn't exist' sort of message
     - 3. I tried looking for a post that did re-redirect but I think that behavior might no longer exist

     since I couldn't find an url that replicated this expected behavior I think
     it's best to remove this, as it really is quite broken and not really testing
     what it was meant to be testing.

     the one thing I'm wondering is if we should take a look at the regex so
     that an url like this would hit item
- test "sets unique title from page description when FB post ID is obscured in URL"
I'm trying to follow a similar logic with all of the tests, so they are clearer
and easier to update
- updated test "should set parser url to full URL when the facebook html og:url is relative"
this one was called "should get canonical URL parsed from facebook html when it is relative"
but I didn't really understand it when looking at the test? so I tried to rename it
also updated it to make sure the final url is the one we want
- test "should return html and empty description when FB url is private"
this one is also not being caught by our set_facebook_privacy_error
     Both of these issues relate to a recent PR we made to try to identify
     login pages https://github.com/meedan/pender/pull/382

**general notes**
- changed all the .to_return(status: 200, body: {}.to_json) to
.to_return(status: 200, body: crowdtangle_response_not_found)
I think this makes what is going on clearer
- though adding the crowdtangle request is not always necessary, I think
adding it makes extra clear the steps that are being followed by the parser
- changed a bunch of WebMock.stub_request(:any to WebMock.stub_request(:get
because we are only stubbing and testing get requests
- Left some notes: in a few tests we are testing Media behavior, so I
wonder if these tests should be here