add test "should update media cache if refresh is requested"
Not too sure about this one. I had run it a bunch of times and it worked,
though I'm not sure about the last three lines. Should they be needed?
After running playing around with it and eventually restoring it, it is
now failing? So at least it seems flaky.