Fixing cache when ID changes
The URL ID, used as a cache key, is calculated in runtime, at different times, when writing and when reading. This means that the cache is written to a key but read from another one. This is a pretty edge case, but this PR turns it safer and more consistent.

Reference: CV2-3749.