update test to use WebMock
I need to work on two tests that broke and I think using WebMock will make
it easier. So I updated this tes (which was passing) to make sure it would work.
I'll keep the WebMock here as well as there was no need for that request
to be live.