update when we archive: fix tests
We only want to archive if a request to refresh the cache is made or if
the data for the url does not yet exist inside the cache.

I missed one condition when updating the when to archive logic.
Sometimes we might have cache, but the data['archives'] might be empty.
In this case, if we are sending the request to archive, we will need
to refresh the cache (so we can update the archive value)

This fixed one test. But there are still two failing tests.