update when we archive: fix tests and refactor
One condition was still not being met: We should refresh the archive
if a new archiver is added.

So I moved all of the conditions to one method that only worries about
if it should be archived or not. I think this is easier to read and we
can see all the conditions at once.

There is still one failing test, but it's related to the video folder,
so I'm not sure it's related to any of these changes.