Update when to archive to archive.org and perma.cc (#406)
* update when we archive

We want to archive if: 
- a request to refresh the cache is made 
- the data for the url is not yet cached
- the cache exists, but the data['archives'] is empty
- the cache exists for one archiver, but the archiver is changed or a new one is added

I moved all of the conditions to one method that only worries about
if it should be archived or not. I think this is easier to read and we
can see all the conditions at once.

* remove logic checking if it should be archived from perma_cc

this should be handled by the parent class. media_perma_cc_archiver
and media_archive_org_archiver should only worry about making the archiving request