update death handler to notify when it's not a RetryLimitHit
Some errors where happening during a delayed job (`get_archive_org_status`),
so our `handle_archiving_exceptions` wasn't handling those.

Meaning they would be retried multiple times without being re-raised
as RetryLater. When they hit the death handler limit_hit_exception would
be nil because they would fail
`if original_exception.is_a?(Pender::Exception::RetryLater)`

We are now handling errors inside `get_archive_org_status`, but we are
adding the else here as an extra safety measure. We want to see what kind
of errors we get for a while and then decide if this i needed or not, and
what the best way to handle this is.