update test "should update media with error when Archive.org can't archive the url"
Since we are now raising Pender::Exception::ArchiveOrgError I had to
update this test. I also re-ordered the tests a bit so it was easier
for me to see which tests we had related to this.

I think I might need to update some tests names, but I'll do this while
working on refactoring the tests (I'm already working on this)