start to refactor tests
- I went through the tests and tried to reorganize them and prepare them
to use our setup and teardown methods

- remove test "MediaArchiver should not notify Sentry when the worker
hits the maximum number of retries" : it helped when I was working on
the duplication of sentry warnings issue, but I don't think we need to
keep this. It has proven to be flaky, and I think it can cause more
confusion than help.

wip: refactor tests