mock perma_cc failed test and fix perma_cc error raising
At first the mocked test seemed to pass, but it was failing because
I wasn't mocking the response correct so response.message was nil.
After I mocked it correctly it stopped adding it to data.

That was happening because we were raising the first error as
RetryLater and we don't set the data when we rescue from RetryLater,
only StandardError (media_archiver.rb:93).

Since I had to change this, I went ahead and created PermaCcError

Note: The same test is run here twice in slightly different ways, I will
delete one