update error handling in perma_cc
While working on refactorinf the tests:
"At first the mocked test seemed to pass, but it was failing because
I wasn't mocking the response correct so response.message was nil.
After I mocked it correctly it stopped adding it to data.

That was happening because we were raising the first error as
RetryLater and we don't set the data when we rescue from RetryLater,
only StandardError (media_archiver.rb:93)."

Besides that, we don't want to retry if the user's perma_cc usage limit
is hit.