3887 – Better error handling in PermaCC (#412)
* add PermaCcError, update error handling in perma_cc and update tests

We want PermaCC to have specific errors, like ArchiveOrg, so we can better track and handle errors. So we created PermaCcError, and it's also using TooManyCaptures once we hit PermaCC's limit.

We want PermaCcError to be retried, update the data and notify the webhook. 
We want TooManyCaptures to NOT be retried ( we don't want to retry if the user's perma_cc usage limit is hit), but still update the data and notify the webhook