add location and error instead os overwriting Everything
Not too sure about this, BUT the issue here is:
- We need to set up the data in media archiver when we handle errors
so that if a request is not completed (a connection error) we still
add error to data
- We need to set up the data inside ArchiveOrg, so that we are able to add
a snapshot if one exists

What was happening:
We would get a ArchiveOrgError, would set the snapshot, but since that
error is raised, we would hit the error handling inside MediaArchiver and
that would overwritte the data we set. We would lose location.

We want to make sure we always add the error if there is one, and we
want to keep the location when we have one.

This only affects archive_org, and not perma_cc (since we don't get a
snapshot there)