add 'X-Priority-Reduced'
Priority reduction

Clients wishing to avoid rate limiting may opt to schedule their change
task at a reduced priority. When the X-Accept-Reduced-Priority request
header is set to a true-ish value (e.g., 1), a client submitting a task
for execution can avoid rate limiting. If the user is being rate limited,
the task will be queued at a reduced priority rather than returning a 429
Too Many Requests.

If priority reduction occurs, the X-Priority-Reduced header is returned
with the 200 OK response. The header value is the task’s reduced priority
number (e.g., -7 or -9).

Clients should be prepared to receive a 429 Too Many Requests even if
the X-Accept-Reduced-Priority header is sent.

https://archive.org/developers/md-write.html#priority-reduction