[wip] 3886 – Clean up archiver tests (#411)
We had a lot of tests that were making real requests that weren't relevant to this test. So we decided to do a refactoring of the tests. What we did:
	1. mock all requests
	2. add new tests when needed
	3. use one setup and teardown for all tests
	4. skip tests related to video-archiving (we don't use this feature anymore, it will be dealt with in another ticket)
	5. general clean up

While working on the tests there were a few issues that appeared, I ended up dealing with them here (they should have been separate tickets/PRs):
	1. skip_perma_cc_archiver should return true if key.nil? (commit)
	2. mock perma_cc failed test and fix perma_cc error raising (commit)
	3. deal with sidekiq warning (commit)
	4. update how we update cache (commit)