update a test
this test worked for me locally but kept failing on Travis. I think
probably because of being IPs from different origins. Both descriptions
do exist, but I was not able to reproduce it locally.

Since we don't have a parser for BBC, testing for a string is enough for
any generic url.