Cv2 3885 remove youtube dl video archiving from pender (#417)
* Remove youtube-dl video archiving

Remove youtube-dl and all its references and tests since it is no
longer supported.

Closes: CV2-3885

* Remove youtube-dl tests from archiver_test.rb

Removing unneeded tests because we are not supporting youtube-dl anymore.

* update a test
this test worked for me locally but kept failing on Travis. I think
probably because of being IPs from different origins. Both descriptions
do exist, but I was not able to reproduce it locally.

Since we don't have a parser for BBC, testing for a string is enough for
any generic url.