3480 – Upgrade to Sidekiq 7 (#416)
1. upgrade Redis (Redis 6.2+ is now required)

2. strict arguments
a. Sidekiq has been logging a warning if JSON-unsafe arguments are passed to perform_async. This warning switched to an exception in Sidekiq 7.0.
b. We could silence strict args, but I was able to fix the instances we had

3. delayed extensions
a. They were disabled in Sidekiq 5 and removed in Sidekiq 7. 
b. I replaced where we were using delayed extensions to use a Job

4. logs
a. Sidekiq info logs were quite noisy when running the tests, so I configured log level to WARN

5. Redis/Travis
a. I ran into some Travis/Redis issues, because of that some tests that depended on Redis were failing in Travis. The solution to that was specifying the distro to jammy inside travis.yml