change trigger from push to pull_request
The github-slack integration is looking for “pull request opened” events
but we had marked the event trigger for the workflow to be on push commits.
they have to match, so we are changing it to the pull_request event trigger
in the test workflow.

We might be able to change it on the slack integration, but I thought it
might be better to keep it the same as presto and timpani.