add on push to master
On timpani and presto we have a separate workflow for the deploy where
we run tests when we push to master.
For pender we don't have a deploy workflow yet, since this is deployed
on gitlab, so we want this to also run when there is a push to master.