4755 – Add otel collector (#456)
This adds otel-collector. It scrapes the Prometheus endpoint for our metrics and sends it to Honeycomb.

I added the bare minimum configuration for it to work. There are features that might be interesting, but I think we could add those as we add the metrics and understand what makes sense for us.

- Adding a batch processor might be interesting: https://github.com/open-telemetry/opentelemetry-collector/blob/main/processor/batchprocessor/README.md
- Maybe activating internal telemetry? https://opentelemetry.io/docs/collector/internal-telemetry#activate-internal-telemetry-in-the-collector

Resources
 - https://docs.honeycomb.io/integrations/metrics/prometheus/
 - https://docs.honeycomb.io/send-data/opentelemetry/collector/
 - https://github.com/open-telemetry/opentelemetry-collector
 - https://opentelemetry.io/docs/collector/quick-start/
 - https://opentelemetry.io/docs/collector/installation/

References: 4755, 4396
PR: 456