add metrics to get more 'specific' data
Sometimes we just want the overall total, without breaking up the data.
This was not so easy to do in the Honeycomb side, so I added more metrics
to make building some of the boards we need easier.

For example:
We were unable to get the total number of items parsed by page, because
it would also break the metric by host. We are using MAX, So if we made
requests for 2 different hosts, it would show 1 parsed by page.

We tried using SUM, but then that counts ALL the events sent. So it
isn't great.