4520 – Add metrics to code and create Honeycomb status dashboard: parser requests [WIP] (#458)
We are now sending metrics related to parsing requests. We are keeping track of: service_name, parser_name, parsed_host and parsing_status.

What we want to answer:
    total of parsing requests per x amount of time
    total of parsing requests per x amount of time grouped by partner [TBD: not possible to do right now]
    total of parsing requests per x amount of time grouped by parser
    total of parsing requests of an unique host, which does not have a dedicated parser, per x amount of time
    total of errors and total of successes of our parsing requests per x amount of time

Why we are sending the data we are sending:

I ended up splitting the metrics so we can get more granular data and more generic data. I did this because it was proving difficult to get the information we wanted on the Honeycomb side.

For some cases we wanted a SUM of the value send by the metric, but I think since the otel-collector sends the scraped data every 15 seconds we end up with a bunch of events. The SUM doesn't really show what we need. We can zoom in into specific points in time, and I have set the default granularity to match the otel-collector. All that helps, but doesn't get 100% what we would like.

I ended up using MAX instead of SUM. But that was an issue when we had more granular metrics and wanted to know, for example, the amount of parsing requests to the page parser, without breaking it up by host.

If we made 2 parsing requests to page but to different hosts, those would be split. And we would get the MAX 1 instead of 2.


References: 4396, 4520
PR: 458