request_url directly to take advantage of error handling
1.
This was doing work that is already handled by request_helper.
The one line I was more cautious about was:
headers = RequestHelper.extended_headers(uri)

But once that seems to only be relevant if it's a twitter parser.

2.
Remove requested_url:
It was sending the url but never using it, instead it needs oembed_url