5215 – Don't use the proxy for Facebook and Instagram links (#492)
Removed proxy usage for Facebook and Instagram:
since we are now using Apify to parse both, we don't need to use the proxy anymore.

Also had to remove a test:
1. The assertion we were making here is not the correct one:
The expected outcome for Pender is to always be able to return data to
the client - because on Check API side we always want the item to be
created. So, if it fails for any reason (and that should be true for
any parser), it should return HTTP 200 and a data structure with at
least the title as the URL and an empty description.

2. We already test this behavior in
rails test test/integration/parsers/instagram_item_test.rb:22

Note: I think it might be a good idea review our controller tests, they
could be redundant, and we should confirm if there are any other wrong
assertions.

References: CV2-5215, CV2-5664
PR: 492